First Commit - I created the first section of the template website
First off, I should have commited more through this
process and in the future I will, but outlined below are the steps taken to
get to this point.

1. I created the index.html, linked it to index-style.css
and began adding divs and other elements that were
requires as per the wireframe.
2. I added an image for the header logo and added the
text for the 3 header links eventhough they do not
go anywhere.
3. I used flexbox to have the header logo by itself
on the lefthand side.
4. I added a slight gap between the header links.
5. Using flexbox I managed to orient the links and
image the same way it is done in the wireframe.
6. I started creating the second half of the "header"
section, I refer to this part as the sign up/image section
7. I added an image to the image section of the webpage.
8. I created the sign up button and the surrounding text
9. I used flexbox to make sure the image and the sign up
both take up the same amount of space.
10. I used flexbox to orient the sign up section into
a column as per the wireframe.
11. I changed fonts and colors to reflect the wireframe accurately.